requirements.txt
apiai==1.2.3
beautifulsoup4==4.9.0
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
convertapi==1.1.1
idna==2.9
numpy==1.18.2
pathlib==1.0.1
Pillow==7.1.1
pycryptodome==3.9.7
pyTelegramBotAPI==3.6.7
requests==2.10.0
six==1.14.0
soupsieve==2.0
tenacity==5.1.5
urllib3==1.25.8
